# docs/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: HaoBlog API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/v1/admin/media/uploads:
    post:
      operationId: createAdminMediaUpload
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MediaUploadRequest' } } }
      responses:
        '201':
          description: Short-lived COS POST policy
          content: { application/json: { schema: { $ref: '#/components/schemas/MediaUploadResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfInvalid' }
        '503': { $ref: '#/components/responses/MediaStorageUnavailable' }
  /api/v1/admin/media/uploads/{uploadId}/complete:
    post:
      operationId: completeAdminMediaUpload
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - { name: uploadId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Confirmed media asset
          content: { application/json: { schema: { $ref: '#/components/schemas/MediaAssetResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfInvalid' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '410': { $ref: '#/components/responses/MediaUploadExpired' }
        '503': { $ref: '#/components/responses/MediaStorageUnavailable' }
  /api/v1/admin/media/{mediaId}:
    delete:
      operationId: recycleAdminMedia
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/CsrfHeader' }
        - { name: mediaId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Media asset marked as deleted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfInvalid' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: HAOBLOG_SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
...
    ResourceNotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
...
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
...
    Unauthenticated:
      description: Authentication required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
...
    CsrfInvalid:
      description: Invalid CSRF token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
...
    MediaUploadExpired:
      description: Upload intent expired
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
    MediaStorageUnavailable:
      description: Object storage metadata service unavailable
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemResponse' }
  schemas:
...
    ProblemResponse:
      type: object
      required: [code, title, detail, traceId]
      properties:
        code: { type: string }
        title: { type: string }
        detail: { type: string }
        traceId: { type: string }
        currentVersion: { type: integer, format: int64, minimum: 0 }
...
    MediaUploadRequest:
      type: object
      required: [mimeType, sizeBytes, width, height, sha256]
      properties:
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 5242880 }
        width: { type: integer, minimum: 1, maximum: 2560 }
        height: { type: integer, minimum: 1, maximum: 2560 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    MediaUploadResponse:
      type: object
      required: [uploadId, objectKey, uploadUrl, fields, expiresAt]
      properties:
        uploadId: { type: string, format: uuid }
        objectKey: { type: string }
        uploadUrl: { type: string, format: uri }
        fields: { type: object, additionalProperties: { type: string } }
        expiresAt: { type: string, format: date-time }
    MediaAssetResponse:
      type: object
      required: [id, objectKey, publicUrl, mimeType, sizeBytes, width, height, sha256, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        objectKey: { type: string }
        publicUrl: { type: string, format: uri }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        sizeBytes: { type: integer, format: int64, minimum: 1 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { type: string, enum: [AVAILABLE, DELETED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...